requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
PyPDF2
tenacity
playwright
requests
bs4
openai
python-dotenv
pymupdf
langchain
langchain_openai
langchain-core
lxml
mysql-connector-python
sqlalchemy
apscheduler
networkx
pandas
numpy
scikit-learn
gensim
node2vec